requirements.txt
# Core ML
torch>=2.1.0
torchaudio>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Audio processing
librosa>=0.10.1
soundfile>=0.12.1

# Generative models (Stable Diffusion)
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0
Pillow>=10.0.0

# Web app
streamlit>=1.30.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
pandas>=2.1.0

# Testing
pytest>=7.4.0

# Development (optional)
jupyter>=1.0.0
